Clean non-applicable SMBIOS info for serial, part, and version matches
When cleaning out well-known bogus values from a DMI contents, this
change adds a filter for values containing the string "applicable" for
the serial number (as well as "part" and "version" matches). For
example, if "Not Applicable" is returned from SMBIOS for the serial
number (DMI string "system-serial-number"), this change will not allow
the string "Not Applicable" from being saved as the grain value.

Signed-off-by: Perry Zipoy <perry.zipoy@ni.com>